requirements.txt
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.7.0
chromadb>=0.4.0
pypdf>=3.8.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
loguru>=0.7.0
numpy>=1.21.0
pandas>=1.3.0
python-dotenv>=1.0.0